4B/Week3 Files/rohan-kothapalli/requirements.txt
# Rohan Kothapalli - Bot Logic Core Requirements
# JKYog WhatsApp Bot - Week 3 Assignment

# OpenAI for intent classification
openai>=1.12.0

# Environment variables
python-dotenv>=1.0.0

# Numerical computing for embeddings
numpy>=1.24.0

# Database (Chakradhar's work)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Google Maps API (Subodh's work)
googlemaps>=4.10.0

# Google Calendar API (Subodh's work)
google-auth>=2.0.0
google-api-python-client>=2.0.0
